Changed PWM settings, because it's hard to get precise load current in 8 bit PWM from ShiftPWM using 595 shift registers. Adjusted current measured with multimiter for each cell.
Added Last state display when battery is pulled out, with option to continue after returning it back (5s delay to clear).

Measured the result capacity and compared to OPUS BT-C3100. In worst case for bad 18650 cells that are under 1000mAh the difference of measurements is >50%. Closer to 2000mAh the difference becomes <5% (~100mAh).

PWM from 595 will be replaced with 16-chanel 12bit PCA9685 modules or with TLC5940 chips